Create notebook
The purpose of the "notebooks" folder is to store Jupyter Notebook files related to the fraud detection project. Jupyter Notebooks provide an interactive and reproducible environment for data analysis, visualization, and model development. This folder will contain notebooks that demonstrate various aspects of the project, including data exploration, feature engineering, model training, evaluation, and visualization. By organizing the notebooks in a dedicated folder, it becomes easier to navigate and manage the project's analysis and documentation.